+++ crates/ruma-macros/src/serde.rs
@@ -3,6 +3,7 @@
 pub mod as_str_as_ref_str;
 pub mod attr;
 pub mod case;
+pub mod debug_as_ref_str;
 pub mod deserialize_from_cow_str;
 pub mod display_as_ref_str;
 pub mod enum_as_ref_str;